% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{adjParams}
\alias{adjParams}
\title{Casemix adjustment}
\usage{
adjParams(model = "logistic", method = "use_all", nfolds = 5L)
}
\arguments{
\item{model}{default is logistic regression}

\item{method}{One of "use_all" or "out_of_fold". Default is "use_all".}

\item{nfolds}{If method is "out_of_fold" how many folds? Default is 5.}
}
\description{
Parameters controlling the casemix adjustment model
}
