% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc_pr}
\alias{auc_pr}
\title{Area under the precision recall curve}
\usage{
auc_pr(prob, truth, length_out = 100)
}
\arguments{
\item{prob}{prediction (probability)}

\item{truth}{observed value}

\item{length_out}{100}
}
\description{
Area under the precision recall curve
}
