% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc_roc}
\alias{auc_roc}
\title{Area under the receiver operating characteristic curve}
\usage{
auc_roc(prob, truth, length_out = 100)
}
\arguments{
\item{prob}{prediction (probability)}

\item{truth}{observed value}

\item{length_out}{100}
}
\description{
Area under the receiver operating characteristic curve
}
