% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{bonferroni}
\alias{bonferroni}
\title{Bonferroni multiplicity adjustment}
\usage{
bonferroni(limits, N, ...)
}
\arguments{
\item{limits}{alpha leves}

\item{N}{number of tests}
}
\description{
Bonferroni multiplicity adjustment
}
