% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{check_distTarget}
\alias{check_distTarget}
\title{Check arguments on distTarget}
\usage{
check_distTarget(x)
}
\arguments{
\item{x}{list created from call to distTarget.}
}
\description{
Checks whether the arguments to distTarget are correct.
}
