% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{check_formula}
\alias{check_formula}
\title{Check the input to funnel}
\usage{
check_formula(formula, var_names)
}
\arguments{
\item{formula}{formula for funnel}

\item{var_names}{names of variables in dataset provided to funnel}
}
\description{
Checks whether the input to the funnel function is correct.
}
