% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{check_pointTarget}
\alias{check_pointTarget}
\title{Check arguments on pointTarget}
\usage{
check_pointTarget(x)

check_pointTarget(x)
}
\arguments{
\item{x}{pointTarget}

\item{x}{distTarget}
}
\description{
Check arguments on pointTarget

Check arguments on distTarget
}
