% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{confusion}
\alias{confusion}
\title{labels data like a confusion matrix}
\usage{
confusion(pred, truth)
}
\arguments{
\item{pred}{prediction (binary)}

\item{truth}{observed value}
}
\description{
labels data like a confusion matrix
}
