% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dispersion}
\alias{dispersion}
\title{Dispersion parameters}
\usage{
dispersion(funnelData, trim = NULL)
}
\arguments{
\item{funnelData}{funnel plot data}

\item{trim}{winsorisation}
}
\description{
Calculate the inflation parameters for situations where overdispersion is present.
}
