% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{distTarget}
\alias{distTarget}
\title{Target is a distribution}
\usage{
distTarget(limits = 0.05, trim = 0, standardised = FALSE)
}
\arguments{
\item{limits}{a numeric vector containing the (1-limit)100\% values for the control limits.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end.}

\item{standardised}{a logical value indicating whether standardise the per cluster outcome. For binary data this is y = observed/expected.}
}
\description{
A description of how the funnel plot target and control limits should be constructed for a distribution target.
}
