% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{eg_sample_func}
\alias{eg_sample_func}
\title{Example sampling function for simulate}
\usage{
eg_sample_func(n)
}
\arguments{
\item{n}{number of sample to take}
}
\description{
Example sampling function for simulate
}
