% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evalCasemixAdj}
\alias{evalCasemixAdj}
\title{Evaluate ability of casemix adjustment variables to predict outcome}
\usage{
evalCasemixAdj(funnelRes, method = "cv", folds = 5L)
}
\arguments{
\item{funnelRes}{a funnel plot object}

\item{method}{"cv" for cross-validation}

\item{folds}{an integer (>= 2) indicating the number of cross validation folds}
}
\description{
Evaluate ability of casemix adjustment variables to predict outcome
}
