% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{fdr}
\alias{fdr}
\title{False discovery rate multiplicity adjustment}
\usage{
fdr(limits, N, pval)
}
\arguments{
\item{limits}{a numeric vector containing the (1-limit)100\% values for the control limits.}

\item{N}{number of clusters (institutions)}

\item{pval}{p-values for FDR calculations}
}
\description{
Adjust control limits to correct for multiple comparisons to a target.
}
