% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{funnel}
\alias{funnel}
\title{Table of results for risk adjusted funnel plots.}
\usage{
funnel(formula, control = pointTarget(), data)
}
\arguments{
\item{formula}{a two-sided formula. Either of the form y ~ x1 + x2 | group,
or if no covariates use y ~ 1 | group}

\item{control}{a description of how the funnel plot target and control limits should be constructed. See the control functions
pointTarget() and distTarget().}

\item{data}{a data frame containing the variables named in formula}
}
\description{
Perform risk adjusted cluster (institution) comparison for the purpose of outlier discovery.
}
\section{Author(s)}{

The package is based on Spiegelhalter (2005) and Jones, Ohlssen & Spiegelhalter (2008). All errors in implementation and are the responsibility of the package authors (Oisin Fitzgerald).
}

\section{References}{

Benjamini, Y., & Hochberg, Y. (1995). Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the royal statistical society. Series B (Methodological), 289-300.

Spiegelhalter, D. J. (2005). Funnel plots for comparing institutional performance. Statistics in medicine, 24(8), 1185-1202.

Jones, H. E., Ohlssen, D. I., & Spiegelhalter, D. J. (2008). Use of the false discovery rate when comparing multiple health care providers. Journal of clinical epidemiology, 61(3), 232-240.
}

