% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{funnelModel}
\alias{funnelModel}
\title{Table of results for risk adjusted funnel plots.}
\usage{
funnelModel(formula, control = pointTarget(), adj = adjParams(), data)
}
\arguments{
\item{formula}{a two-sided formula. Either of the form y ~ x1 + x2 | group,
or if no covariates use y ~ 1 | group}

\item{control}{a description of how the funnel plot target and control limits should be constructed. See the control functions
pointTarget() and distTarget().}

\item{adj}{a description of the algorithm that will be used in calculating the expected outcome for each observation for the purpose of casemix adjustment.
See adjModel() for more details.}

\item{data}{a data frame containing the variables named in formula}
}
\value{

}
\description{
Perform risk adjusted cluster (institution) comparison for the purpose of outlier discovery.
}
\section{Author(s)}{

The package is based on Spiegelhalter (2005) and Jones, Ohlssen & Spiegelhalter (2008). All errors in implementation and are the responsibility of the package author (Oisin Fitzgerald).
}

