% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_example}
\alias{generate_example}
\title{Simulate data for a funnel plot}
\usage{
generate_example(N, n_min, n_max, mean, sd = 0, sample_func = NULL)
}
\arguments{
\item{N}{number of clusters (institutions)}

\item{n_min}{limit on the number of observations within the smallest cluster (institution)}

\item{n_max}{limit on the  of observations within the largest cluster (institution)}

\item{mean}{mean of cluster (institution) mean distribution}

\item{sd}{standard deviation of cluster (institution) mean distribution}

\item{sample_func}{function to sample jth cluster (institution) size in a form other than Unif(n_min,n_max)}
}
\description{
Simulate data for a funnel plot
}
