% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getFunnelFormula}
\alias{getFunnelFormula}
\title{Edit funnel plot formula}
\usage{
getFunnelFormula(formula)
}
\arguments{
\item{formula}{formula passed to funnel.}
}
\description{
Edit formula passed to funnel to enable call to stats::glm
}
