% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funnel_plot.R
\name{grouped_proportion}
\alias{grouped_proportion}
\title{Calculate proportion of}
\usage{
grouped_proportion(observed, group, expected = NULL, target = NULL)
}
\description{
Calculate proportion of
}
