% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{groupOutcomes}
\alias{groupOutcomes}
\title{Calculate proportion of ???}
\usage{
groupOutcomes(observed, expected, id, target = NULL)
}
\arguments{
\item{observed}{observed outcome (a vector)}

\item{expected}{expected outcome (a vector)}

\item{id}{institution identifier}

\item{target}{institution target value (a proportion)}
}
\description{
Calculate proportion of ???
}
