% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{normHypTest}
\alias{normHypTest}
\title{Normal assuming hypothesis test on grouped data}
\usage{
normHypTest(obs, exp, stdErr)
}
\arguments{
\item{obs}{observed}

\item{exp}{null hypothesis}

\item{stdErr}{standard error}
}
\description{
Compare institutions to a target using a normality assumption.
}
