% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{outliers}
\alias{outliers}
\title{List institutional outliers}
\usage{
outliers(funnelRes, limits = NULL)
}
\arguments{
\item{funnelRes}{funnel plot object}

\item{limits}{Optional argument indicating which subset of the control limits should be used for selecting outlying institutions.}
}
\description{
Return a ?type vector of all outliers
}
