% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.fsim}
\alias{plot.fsim}
\title{Plot the simulated data}
\usage{
\method{plot}{fsim}(x, effect = "true", method = "dist", ...)
}
\arguments{
\item{x}{the simulation results}

\item{effect}{whether to plot the "true" or "estimated" cluster (institution) outcome value}

\item{method}{whether to assume a "point" or "distribution" target}

\item{...}{other arguments to plot.fsim}
}
\description{
Plot the simulated data
}
