% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.funnelRes}
\alias{plot.funnelRes}
\title{plot funnel plot}
\usage{
\method{plot}{funnelRes}(funnelRes, identify = "all", label = "none",
  lengthOut = 500, ...)
}
\arguments{
\item{funnelRes}{funnel plot object}

\item{identify}{all}

\item{label}{none}

\item{lengthOut}{500}
}
\description{
plot funnel plot
}
