% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLimits.funnelRes}
\alias{plotLimits.funnelRes}
\title{helper function for plotting the funnel plots}
\usage{
\method{plotLimits}{funnelRes}(funnelRes, lengthOut)
}
\arguments{
\item{funnelRes}{funnel plot object}

\item{lengthOut}{500}
}
\description{
helper function for plotting the funnel plots
}
