% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLimits}
\alias{plotLimits}
\title{helper function for plotting the funnel plots}
\usage{
plotLimits(funnelRes, lengthOut)
}
\arguments{
\item{funnelRes}{funnel plot object}

\item{lengthOut}{resolution of control limits. Number of points sampled to construct limit lines. default = 500.}
}
\description{
Determines the position of the control limits at N values.
}
