% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{pointLimits}
\alias{pointLimits}
\title{Calculate control limits}
\usage{
pointLimits(target, n, N, inflationFactor, control, pval)
}
\arguments{
\item{target}{institution target value (a proportion)}

\item{n}{precision values at which to calculate limit (vector)}

\item{N}{number of clusters (institutions)}

\item{inflationFactor}{Unexplained variation}

\item{control}{description of task}

\item{pval}{p-values for FDR calculations}
}
\description{
Calculate the control limits for use in plotting and comparison of individual clusters (institutions) to a point target.
}
