% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ppv}
\alias{ppv}
\title{precision, positive predictive value}
\usage{
ppv(pred, truth)
}
\arguments{
\item{pred}{prediction (binary)}

\item{truth}{observed value}
}
\description{
precision, positive predictive value
}
