% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pr}
\alias{pr}
\title{precision recall curves}
\usage{
pr(prob, truth, length_out = 100)
}
\arguments{
\item{prob}{prediction (probability)}

\item{truth}{observed value}

\item{length_out}{100}
}
\description{
precision recall curves
}
