% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.funnelRes}
\alias{print.funnelRes}
\title{print function}
\usage{
\method{print}{funnelRes}(x, ...)
}
\arguments{
\item{x}{funnel plot object.}

\item{...}{other arguments to print.}
}
\description{
print function
}
