% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{randomEffectLimits}
\alias{randomEffectLimits}
\title{Calculate ...}
\usage{
randomEffectLimits(target, n, effectVar, control)
}
\arguments{
\item{target}{...}

\item{control}{}
}
\description{
Calculate ...
}
