% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_means}
\alias{sample_means}
\title{Sample means}
\usage{
sample_means(N, mean, sd)
}
\arguments{
\item{N}{number of clusters (institutions)}

\item{mean}{mean of cluster (institution) mean distribution}

\item{sd}{standard deviation of cluster (institution) mean distribution}
}
\description{
Sample means
}
