% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_nobs}
\alias{sample_nobs}
\title{Sample a certain number of observations}
\usage{
sample_nobs(N, n_min, n_max, sample_func)
}
\arguments{
\item{N}{number of clusters (institutions)}

\item{n_min}{limit on the number of observations within the smallest cluster (institution)}

\item{n_max}{limit on the  of observations within the largest cluster (institution)}

\item{sample_func}{function to sample jth cluster (institution) size in a form other than Unif(n_min,n_max)}
}
\description{
Sample a certain number of observations
}
