% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tpr}
\alias{tpr}
\title{recall, sensitivity, true positive rate}
\usage{
tpr(pred, truth)
}
\arguments{
\item{pred}{prediction (binary)}

\item{truth}{observed value}
}
\description{
recall, sensitivity, true positive rate
}
