% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{bonferroni}
\alias{bonferroni}
\title{Bonferroni multiplicity adjustment.}
\usage{
bonferroni(limits, N, ...)
}
\arguments{
\item{limits}{alpha levels}

\item{N}{number of tests}

\item{...}{additional args}
}
\description{
Adjusts control limits to correct for multiple comparisons to a target.
}
