% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{calcExpected}
\alias{calcExpected}
\title{Casemix adjustment}
\usage{
calcExpected(formula, data, params)
}
\arguments{
\item{formula}{model formula in the form outcome ~ covariates}

\item{data}{dataset}

\item{params}{adjParams}
}
\description{
Casemix adjustment
}
