% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{check_formula}
\alias{check_formula}
\title{Check the input to funnelplot::funnel}
\usage{
check_formula(x, var_names)
}
\arguments{
\item{x}{formula for funnel}
}
\description{
Check the input to funnelplot::funnel
}
