% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{check_pointTarget}
\alias{check_pointTarget}
\title{Check arguments on pointTarget}
\usage{
check_pointTarget(x)
}
\arguments{
\item{x}{list created from call to pointTarget.}
}
\description{
Checks whether the arguments to pointTarget are correct.
}
