% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{contAdjustBin}
\alias{contAdjustBin}
\title{Continuity adjusted binomial limits}
\usage{
contAdjustBin(limit, n, target)
}
\arguments{
\item{limit}{control limit}

\item{n}{precision values at which to calculate limit}

\item{target}{institution target value}
}
\description{
Continuity adjusted binomial limits
}
