% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{distTarget}
\alias{distTarget}
\title{Target is a distribution}
\usage{
distTarget(limits = 0.05, trim = 0)
}
\arguments{
\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end}
}
\description{
Target is a distribution
}
