% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evalCasemixAdj}
\alias{evalCasemixAdj}
\title{Casemix adjustment}
\usage{
evalCasemixAdj(funnelRes, method = "cv", folds = 10)
}
\arguments{
\item{funnelRes}{funnel plot object}

\item{method}{cross-validation}

\item{folds}{10}
}
\description{
Casemix adjustment
}
