% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{fdr}
\alias{fdr}
\title{False discovery rate multiplicity adjustment}
\usage{
fdr(limits, N, pval)
}
\arguments{
\item{N}{number of institutions}

\item{pval}{p-values for FDR calculations}
}
\description{
False discovery rate multiplicity adjustment
}
