% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{funnel}
\alias{funnel}
\title{Perform risk adjusted institution comparison}
\usage{
funnel(formula, control = pointTarget(), data)
}
\arguments{
\item{formula}{a two-sided formula. Either of the form y ~ x1 + x2 | group,
or if no covariates use y ~ 1 | group}

\item{data}{a data frame containing the variables named in formula

CHANGE normal_approx TO METHOD = "exact", "normal" AND SO ON!}

\item{target}{a list of values that define how the funnel plot should be constructed.
See pointTarget and distTarget.}
}
\description{
Perform risk adjusted institution comparison
}
