% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{funnel}
\alias{funnel}
\title{Table of results for risk adjusted funnel plots.}
\usage{
funnel(formula, control = pointTarget(), data)
}
\arguments{
\item{formula}{a two-sided formula. Either of the form y ~ x1 + x2 | group,
or if no covariates use y ~ 1 | group}

\item{control}{a description of how the funnel plot target and control limits should be constructed. See the control functions
pointTarget() and distTarget().}

\item{data}{a data frame containing the variables named in formula}
}
\description{
Perform risk adjusted cluster (institution) comparison for the purpose of outlier discovery.
}
\section{Author(s)}{

The package is based on Spiegelhalter (2005) and Jones, Ohlssen & Spiegelhalter (2008). All errors in implementation and are the responsibility of the package authors (Oisin Fitzgerald).
}

