% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funnel_plot.R
\name{funnel}
\alias{funnel}
\title{Create the funnel plot}
\usage{
funnel(outcome, group, limits = 0.05, normal_approx = FALSE,
  casemix_adj = NULL, w = NULL, target = NULL,
  ctrl_over_disp = TRUE)
}
\description{
Create the funnel plot
}
