% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{groupOutcomes}
\alias{groupOutcomes}
\title{Calculate summary values per cluster}
\usage{
groupOutcomes(observed, expected, id, target = NULL)
}
\arguments{
\item{observed}{observed outcome (a vector)}

\item{expected}{expected outcome (a vector)}

\item{id}{cluster (institution) identifier}

\item{target}{institution target value (a proportion)}
}
\description{
Calculates summary measures of per cluster (institution) performance
}
