% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.fsim}
\alias{plot.fsim}
\title{Plot the simulated data}
\usage{
\method{plot}{fsim}(sim, effect = "true", method = "dist")
}
\arguments{
\item{sim}{the simulation results}

\item{effect}{true}

\item{method}{dist}
}
\description{
Plot the simulated data
}
