% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.funnelRes}
\alias{plot.funnelRes}
\title{Graph funnel plot}
\usage{
\method{plot}{funnelRes}(x, identify = "all", label = "none",
  lengthOut = 500, ...)
}
\arguments{
\item{x}{funnel plot object}

\item{identify}{show on plot. default = all}

\item{label}{label on plot. default = none}

\item{lengthOut}{resolution of control limits. Number of points sampled to construct limit lines. default = 500.}

\item{...}{Other arguments to plot.funnelRes.}
}
\description{
Produces a funnel plot using ggplot2.
}
