% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{pointLimits}
\alias{pointLimits}
\title{Calculate control limits for pointTarget method}
\usage{
pointLimits(target, n, N, inflationFactor, control, pval)
}
\arguments{
\item{target}{institution target value (a proportion)}

\item{n}{precision values at which to calculate limit}

\item{N}{number of institutions}

\item{inflationFactor}{Unexplained variation}

\item{control}{description of task}

\item{pval}{p-values for FDR calculations}
}
\description{
Calculate control limits for pointTarget method
}
