% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{randomEffectLimits}
\alias{randomEffectLimits}
\title{Calculate control limits}
\usage{
randomEffectLimits(target, n, effectVar, control)
}
\arguments{
\item{target}{cluster (institution) mean distribution mean target value}

\item{n}{precision values at which to calculate limit (vector)}

\item{effectVar}{cluster (institution) mean distribution variance}

\item{control}{description of task}
}
\description{
Calculate the control limits for use in plotting and comparison of individual clusters (institutions) to a distribution target.
}
