% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{randomEffectLimits}
\alias{randomEffectLimits}
\title{Calculate ...}
\usage{
randomEffectLimits(target, n, effectVar, control)
}
\arguments{
\item{target}{institution mean distribution mean target value}

\item{n}{precision values at which to calculate limit}

\item{effectVar}{institution mean distribution variance}

\item{control}{description of task}
}
\description{
Calculate ...
}
