% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.funnelRes}
\alias{summary.funnelRes}
\title{summary function}
\usage{
\method{summary}{funnelRes}(object, ...)
}
\arguments{
\item{object}{funnel plot object.}

\item{...}{other arguments to print.}
}
\description{
summary function
}
